UPD: move get_true_label_cumsum_proba to class method
